+++ Radio/FilterBuilder.cs
@@ -16,84 +16,84 @@ namespace SDRSharp.Radio
     {
         public const int DefaultFilterOrder = 500;
 
-        public static double[] MakeWindow(WindowType windowType, int length)
+        public static float[] MakeWindow(WindowType windowType, int length)
         {
-            var w = new double[length];
+            var w = new float[length];
             for (var i = 0; i < length; i++)
             {
-                double n;
-                double a0;
-                double a1;
-                double a2;
-                double a3;
-                double alpha;
+                float n;
+                float a0;
+                float a1;
+                float a2;
+                float a3;
+                float alpha;
 
-                w[i] = 1.0;
+                w[i] = 1.0f;
                 
                 switch (windowType)
                 {
                     case WindowType.Hamming:
-                        a0 = 0.54;
-                        a1 = 0.46;
-                        a2 = 0.0;
-                        a3 = 0.0;
+                        a0 = 0.54f;
+                        a1 = 0.46f;
+                        a2 = 0.0f;
+                        a3 = 0.0f;
                         w[i] *= a0
-                              - a1 * Math.Cos(2.0 * Math.PI * i / length)
-                              + a2 * Math.Cos(4.0 * Math.PI * i / length)
-                              - a3 * Math.Cos(6.0 * Math.PI * i / length);
+                              - a1 * (float) Math.Cos(2.0 * Math.PI * i / length)
+                              + a2 * (float) Math.Cos(4.0 * Math.PI * i / length)
+                              - a3 * (float) Math.Cos(6.0 * Math.PI * i / length);
                         break;
 
                     case WindowType.Blackman:
-                        a0 = 0.42;
-                        a1 = 0.5;
-                        a2 = 0.08;
-                        a3 = 0.0;
+                        a0 = 0.42f;
+                        a1 = 0.5f;
+                        a2 = 0.08f;
+                        a3 = 0.0f;
                         w[i] *= a0
-                              - a1 * Math.Cos(2.0 * Math.PI * i / length)
-                              + a2 * Math.Cos(4.0 * Math.PI * i / length)
-                              - a3 * Math.Cos(6.0 * Math.PI * i / length);
+                              - a1 * (float) Math.Cos(2.0 * Math.PI * i / length)
+                              + a2 * (float) Math.Cos(4.0 * Math.PI * i / length)
+                              - a3 * (float) Math.Cos(6.0 * Math.PI * i / length);
                         break;
 
                     case WindowType.BlackmanHarris:
-                        a0 = 0.35875;
-                        a1 = 0.48829;
-                        a2 = 0.14128;
-                        a3 = 0.01168;
+                        a0 = 0.35875f;
+                        a1 = 0.48829f;
+                        a2 = 0.14128f;
+                        a3 = 0.01168f;
                         w[i] *= a0
-                              - a1 * Math.Cos(2.0 * Math.PI * i / length)
-                              + a2 * Math.Cos(4.0 * Math.PI * i / length)
-                              - a3 * Math.Cos(6.0 * Math.PI * i / length);
+                              - a1 * (float) Math.Cos(2.0 * Math.PI * i / length)
+                              + a2 * (float) Math.Cos(4.0 * Math.PI * i / length)
+                              - a3 * (float) Math.Cos(6.0 * Math.PI * i / length);
                         break;
 
                     case WindowType.HannPoisson:
-                        n = i - length / 2.0;
-                        alpha = 0.005;
-                        w[i] *= 0.5 * (1.0 + Math.Cos(2.0 * Math.PI * n / length)) * Math.Exp(-2.0 * alpha * Math.Abs(n) / length);
+                        n = i - length / 2.0f;
+                        alpha = 0.005f;
+                        w[i] *= 0.5f * (float) ((1.0 + Math.Cos(2.0 * Math.PI * n / length)) * Math.Exp(-2.0 * alpha * Math.Abs(n) / length));
                         break;
 
                     case WindowType.Youssef:
-                        a0 = 0.35875;
-                        a1 = 0.48829;
-                        a2 = 0.14128;
-                        a3 = 0.01168;
-                        n = i - length / 2.0;
-                        alpha = 0.005;
+                        a0 = 0.35875f;
+                        a1 = 0.48829f;
+                        a2 = 0.14128f;
+                        a3 = 0.01168f;
+                        n = i - length / 2.0f;
+                        alpha = 0.005f;
                         w[i] *= a0
-                              - a1 * Math.Cos(2.0 * Math.PI * i / length)
-                              + a2 * Math.Cos(4.0 * Math.PI * i / length)
-                              - a3 * Math.Cos(6.0 * Math.PI * i / length);
-                        w[i] *= Math.Exp(-2.0 * alpha * Math.Abs(n) / length);
+                              - a1 * (float) Math.Cos(2.0 * Math.PI * i / length)
+                              + a2 * (float) Math.Cos(4.0 * Math.PI * i / length)
+                              - a3 * (float) Math.Cos(6.0 * Math.PI * i / length);
+                        w[i] *= (float) Math.Exp(-2.0 * alpha * Math.Abs(n) / length);
                         break;
                 }
             }
             return w;
         }
 
-        public static double[] MakeLowPassKernel(int sampleRate, int filterOrder, int cutoffFrequency, WindowType windowType)
+        public static float[] MakeLowPassKernel(int sampleRate, int filterOrder, int cutoffFrequency, WindowType windowType)
         {
-            var cutoffRadians = 2 * Math.PI * cutoffFrequency / sampleRate;
+            var cutoffRadians = (float) (2 * Math.PI * cutoffFrequency / sampleRate);
             
-            var h = new double[filterOrder + 1];
+            var h = new float[filterOrder + 1];
             var w = MakeWindow(windowType, filterOrder + 1);
 
             for (var i = 0; i <= filterOrder; i++)
@@ -105,7 +105,7 @@ namespace SDRSharp.Radio
                 }
                 else
                 {
-                    h[i] = Math.Sin(cutoffRadians * n) / n;
+                    h[i] = (float) (Math.Sin(cutoffRadians * n) / n);
                     h[i] *= w[i];
                 }
             }
@@ -113,12 +113,12 @@ namespace SDRSharp.Radio
             return h;
         }
 
-        public static double[] MakeHighPassKernel(int sampleRate, int filterOrder, int cutoffFrequency, WindowType windowType)
+        public static float[] MakeHighPassKernel(int sampleRate, int filterOrder, int cutoffFrequency, WindowType windowType)
         {
             return InvertSpectrum(MakeLowPassKernel(sampleRate, filterOrder, cutoffFrequency, windowType));
         }
 
-        public static double[] MakeBandPassKernel(int sampleRate, int filterOrder, int cutoff1, int cutoff2, WindowType windowType)
+        public static float[] MakeBandPassKernel(int sampleRate, int filterOrder, int cutoff1, int cutoff2, WindowType windowType)
         {
             var bw = (cutoff2 - cutoff1) / 2;
             var fshift = cutoff2 - bw;
@@ -129,17 +129,17 @@ namespace SDRSharp.Radio
             for (var i = 0; i < h.Length; i++)
             {
                 var n = i - filterOrder / 2;
-                h[i] *= 2 * Math.Cos(shiftRadians * n);
+                h[i] *= (float) (2 * Math.Cos(shiftRadians * n));
             }
             return h;
         }
 
         #region Utility functions
 
-        private static void Normalize(double[] h)
+        private static void Normalize(float[] h)
         {
             // Normalize the filter kernel for unity gain at DC
-            var sum = 0.0;
+            var sum = 0.0f;
             for (var i = 0; i < h.Length; i++)
             {
                 sum += h[i];
@@ -153,13 +153,13 @@ namespace SDRSharp.Radio
         // See the bottom of
         // http://www.dspguide.com/ch14/4.htm
         // for an explanation of spectral inversion
-        private static double[] InvertSpectrum(double[] h)
+        private static float[] InvertSpectrum(float[] h)
         {
             for (var i = 0; i < h.Length; i++)
             {
                 h[i] = -h[i];
             }
-            h[(h.Length - 1) / 2] += 1.0;
+            h[(h.Length - 1) / 2] += 1.0f;
             return h;
         }
